requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg[binary]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
bcrypt==4.0.1
pytest
